Resolve parsing ambiguity
This commit includes the removal of support for forward declaration of
enums. This is fine because forward declaration of an enum is not valid
according to the C standard. This is reasonable, because if an
underlying type is not specified for an enum, then the compilier is free
to choose a size based on the number of options in the enum.