Use ALL logics in SMT2 solvers and fix Bool/_ BitVec 1 in cvc4
This changes the default logic in all SMT2 solvers to use the
ALL logic instead of the QF_AUFBV logic. This may have side
effects on performance, and can lead to z3 producing an
"unknown" result iunstead of "sat" or "unsat". However, this
allows cvc4 to produce more results, which is what we care about
here.

Also this changes the way Arrays of Bools are handled by cvc4, to
allow them to use "Bool" instead of "(_ BitVec 1)" which fixes
a parse error in cvc4.