Add handling of unknown result from solver
This handles an "unknown" result from an SMT solver as an "unsat"
result. Preliminary testing shows this does not break anything, but
it is unsound in general.

This fixes potentialy problems with the earlier commits here that
allow solvers to use ALL logics and thus produce an "unknown"
result. So while the overall branch has some limitations, this
commit changes from:
  error when solver returns "unknown"
to instead
  treat "unknown" as "unsat" and continue
which is less sound, but gives more results.