Deduplicate output streams in `smt2_convt::write_footer()`
This avoids bugs arising from some of the member functions of
`smt2::convt` writing to `os` and some of the them writing to `out`.
This was causing some of the generated text to be written out when using
`--outfile`, but not without this argument, in the case where cbmc uses
a temp file to send the formula to the solver itself.

It might be easier to understand `smt2_convt` if the output stream was
passed as an argument to the member functions, instead of kept as a
member variable. This commit replaces the arguments with references to
the member variable because it is a less wide sweeping change and it
still solves the bugs.