Use standardised checks for kinds of expression
The `can_cast_expr` functions are the standardised implementation of the
checking if an expression is a particular sub type of expression. Using
them helps the reader to locate the associated class and its
documentation.

Note that there doesn't seem to be a class associated to `ID_ptrmember`
and it would help readability if such a class were to be created,
documented and used instead of referencing the id directly.